getserver status function move from hostTool to clc/view.
